Use the same-type variable in a loop, as in loop condition
This fixes (most of the) warnings like this:
darktable/src/common/cache.c:382:17: warning: comparison of integers of different signs: 'int' and 'uint32_t' (aka 'unsigned int') [-Wsign-compare]
  for(int k=0; k<=cache->segment_mask; k++)
               ~^ ~~~~~~~~~~~~~~~~~~~

Conflicts:
	src/common/imageio_gm.c
	src/common/imageio_png.c
	src/common/imageio_tiff.c
	src/common/opencl.c
	src/develop/pixelpipe_hb.c
	src/imageio/format/j2k.c
	src/libs/collect.c

+++ src/common/image.c
@@ -537,7 +537,7 @@ uint32_t dt_image_import(const int32_t film_id, const char *filename, gboolean o
 
   if (!glob(pattern, 0, NULL, globbuf))
   {
-    for (int i=0; i < globbuf->gl_pathc; i++)
+    for (size_t i=0; i < globbuf->gl_pathc; i++)
     {
       int newid = -1;
       newid = dt_image_duplicate(id);
@@ -847,7 +847,7 @@ void dt_image_synch_all_xmp(const gchar *pathname)
 
     if (!glob(pattern, 0, NULL, globbuf))
     {
-      for (int i=0; i < globbuf->gl_pathc; i++)
+      for (size_t i=0; i < globbuf->gl_pathc; i++)
       {
         (void)g_unlink(globbuf->gl_pathv[i]);
       }
